VBoxManage: build fix: limit keyboardputscancode 'typing' rate. Simple commands to 'type' ASCII chars.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@64907 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManageControlVM.cpp
@@ -32,6 +32,7 @@
 #include <iprt/getopt.h>
 #include <iprt/stream.h>
 #include <iprt/string.h>
+#include <iprt/thread.h>
 #include <iprt/uuid.h>
 #include <iprt/file.h>
 #include <VBox/log.h>
@@ -228,7 +229,7 @@ static HRESULT keyboardPutScancodes(IKeyboard *pKeyboard, const std::list<LONG> 
 
 #if 1
     HRESULT rc = S_OK;
-    int i;
+    size_t i;
     for (i = 0; i < saScancodes.size(); ++i)
     {
         rc = pKeyboard->PutScancode(saScancodes[i]);
